+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * Address upcoming [Trakt pagination requirements](https://github.com/trakt/trakt-api/discussions/775) for watched methods:
   * Add paginated variants of `Sync.watchedMovies()`, `Sync.watchedShows()`, `Users.watchedMovies()` and 
     `Users.watchedShows()` with required `page` and `limit` parameters. The original methods are deprecated.
-* Add `Sync.watchedEpisodes()` 
+* Add `min` variants of watched methods: `Users.watchedMoviesMin()`, `Users.watchedShowsMin()`, `Sync.watchedMoviesMin()` and `Sync.watchedShowsMin()`
+* Add episode variant of watched methods: `Sync.watchedEpisodes()` and `Sync.watchedEpisodesMin()` 
 
 ## 6.18.0 - 2026-03-06